# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Air Pollution Prediction API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/stations": {
      "get": {
        "tags": [
          "stations"
        ],
        "summary": "List Stations",
        "description": "Stations with a current forecast come first: the upstream sensor feed\nlags for most of the network, so most stations cannot be forecast right now\nand clients need to tell them apart.",
        "operationId": "list_stations_api_v1_stations_get",
        "parameters": [
          {
            "name": "region_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Region Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/StationOut"
                  },
                  "title": "Response List Stations Api V1 Stations Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/stations/{station_id}": {
      "get": {
        "tags": [
          "stations"
        ],
        "summary": "Get Station",
        "operationId": "get_station_api_v1_stations__station_id__get",
        "parameters": [
          {
            "name": "station_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Station Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StationDetailOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/forecast/{station_id}/exceedance": {
      "get": {
        "tags": [
          "exceedance"
        ],
        "summary": "Get Exceedance Summary",
        "operationId": "get_exceedance_summary_api_v1_forecast__station_id__exceedance_get",
        "parameters": [
          {
            "name": "station_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Station Id"
            }
          },
          {
            "name": "pollutant",
            "in": "query",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/Pollutant",
              "default": "pm25"
            }
          },
          {
            "name": "days_ahead",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 14,
              "minimum": 1,
              "default": 5,
              "title": "Days Ahead"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExceedanceSummaryOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ExceedanceDayOut": {
        "properties": {
          "date": {
            "type": "string",
            "format": "date",
            "title": "Date"
          },
          "exceedance_flag": {
            "type": "boolean",
            "title": "Exceedance Flag"
          },
          "exceedance_probability": {
            "type": "number",
            "title": "Exceedance Probability"
          },
          "worst_case_value": {
            "type": "number",
            "title": "Worst Case Value"
          },
          "aqi_category": {
            "type": "string",
            "title": "Aqi Category"
          }
        },
        "type": "object",
        "required": [
          "date",
          "exceedance_flag",
          "exceedance_probability",
          "worst_case_value",
          "aqi_category"
        ],
        "title": "ExceedanceDayOut"
      },
      "ExceedanceSummaryOut": {
        "properties": {
          "station_id": {
            "type": "string",
            "title": "Station Id"
          },
          "pollutant": {
            "type": "string",
            "title": "Pollutant"
          },
          "timezone": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Timezone"
          },
          "days": {
            "items": {
              "$ref": "#/components/schemas/ExceedanceDayOut"
            },
            "type": "array",
            "title": "Days"
          },
          "overall_recommendation": {
            "type": "string",
            "title": "Overall Recommendation"
          }
        },
        "type": "object",
        "required": [
          "station_id",
          "pollutant",
          "days",
          "overall_recommendation"
        ],
        "title": "ExceedanceSummaryOut"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "Pollutant": {
        "type": "string",
        "enum": [
          "pm25",
          "no2"
        ],
        "title": "Pollutant"
      },
...
      "StationDetailOut": {
        "properties": {
          "station_id": {
            "type": "string",
            "title": "Station Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "lat": {
            "type": "number",
            "title": "Lat"
          },
          "lon": {
            "type": "number",
            "title": "Lon"
          },
          "city": {
            "type": "string",
            "title": "City"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active"
          },
          "region_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Region Id"
          },
          "latest_observed_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latest Observed At"
          },
          "has_current_forecast": {
            "type": "boolean",
            "title": "Has Current Forecast",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "station_id",
          "name",
          "lat",
          "lon",
          "city",
          "is_active"
        ],
        "title": "StationDetailOut"
      },
      "StationOut": {
        "properties": {
          "station_id": {
            "type": "string",
            "title": "Station Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "lat": {
            "type": "number",
            "title": "Lat"
          },
          "lon": {
            "type": "number",
            "title": "Lon"
          },
          "city": {
            "type": "string",
            "title": "City"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active"
          },
          "region_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Region Id"
          },
          "latest_observed_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latest Observed At"
          },
          "has_current_forecast": {
            "type": "boolean",
            "title": "Has Current Forecast",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "station_id",
          "name",
          "lat",
          "lon",
          "city",
          "is_active"
        ],
        "title": "StationOut"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}